Update uno.sql for plus cards

+++ db/uno.sql
@@ -3,7 +3,7 @@
 -- https://www.phpmyadmin.net/
 --
 -- Φιλοξενητής: 127.0.0.1
--- Χρόνος δημιουργίας: 24 Δεκ 2019 στις 16:46:49
+-- Χρόνος δημιουργίας: 25 Δεκ 2019 στις 17:08:41
 -- Έκδοση διακομιστή: 10.4.8-MariaDB
 -- Έκδοση PHP: 7.3.11
 
@@ -21,18 +21,31 @@ SET time_zone = "+00:00";
 --
 -- Βάση δεδομένων: `uno`
 --
+CREATE DATABASE IF NOT EXISTS `uno` DEFAULT CHARACTER SET utf8 COLLATE utf8_bin;
+USE `uno`;
 
 DELIMITER $$
 --
 -- Διαδικασίες
 --
+DROP PROCEDURE IF EXISTS `check_do_move`$$
+CREATE DEFINER=`root`@`localhost` PROCEDURE `check_do_move` (IN `table_card_color` ENUM('R','Y','B','G','W'), IN `table_card_symbol` ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N'), IN `c_code` VARCHAR(3))  BEGIN
+	DECLARE plus SMALLINT;
+    SELECT plus_counter INTO plus FROM game_status LIMIT 1;
+    IF (table_card_symbol = '+4' OR table_card_symbol = '+2') AND plus != 0 THEN
+    	CALL plus_do_move(table_card_color, table_card_symbol, c_code);
+    ELSE
+    	CALL do_move(table_card_color, table_card_symbol, c_code);
+    END IF;
+END$$
+
 DROP PROCEDURE IF EXISTS `clean_deck`$$
 CREATE DEFINER=`root`@`localhost` PROCEDURE `clean_deck` ()  BEGIN
 	REPLACE INTO remaining_deck SELECT * FROM deck;
     DELETE FROM hand;
     DELETE FROM table_deck;
     UPDATE player SET username=null, token=null, uno_status= 'not active';
-    UPDATE game_status SET status='not active', p_turn=null, result=null;
+    UPDATE game_status SET status='not active', p_turn=null, result=null, draw_counter=0, plus_counter=0;
 END$$
 
 DROP PROCEDURE IF EXISTS `do_move`$$
@@ -41,34 +54,56 @@ CREATE DEFINER=`root`@`localhost` PROCEDURE `do_move` (IN `table_card_color` ENU
 	DECLARE c_color ENUM('R','Y','B','G','W');
     DECLARE player_turn ENUM('p1', 'p2');
     DECLARE c_symbol ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N');
-    SELECT p_turn INTO player_turn FROM game_status;
-    SELECT d.card_color, d.card_symbol, h.card_id into c_color, c_symbol, c_id FROM hand h INNER JOIN deck d on h.card_id=d.card_id WHERE d.card_code=c_code AND h.player_name=player_turn LIMIT 1;    
-    IF table_card_color = 'W' THEN 
-    	SELECT wild_card_color INTO table_card_color FROM table_deck order by table_id desc limit 1; 
-    END IF;
-    IF c_symbol = 'S' THEN 
-    	CALL skip_do_move(table_card_color, table_card_symbol, c_code, c_color, c_id, player_turn);
-    ELSEIF c_symbol = 'R' THEN
-    	CALL reverse_do_move(table_card_color, table_card_symbol, c_code, c_color, c_id, player_turn);
-    ELSE
-    	CALL regular_do_move(table_card_color, c_code, c_color, c_id, player_turn, table_card_symbol, c_symbol);
-    END IF;
+
+        SELECT p_turn INTO player_turn FROM game_status;
+        SELECT d.card_color, d.card_symbol, h.card_id into c_color, c_symbol, c_id FROM hand h INNER JOIN deck d on h.card_id=d.card_id WHERE d.card_code=c_code AND h.player_name=player_turn LIMIT 1;  
+        IF table_card_color = 'W' THEN 
+            SELECT wild_card_color INTO table_card_color FROM table_deck order by table_id desc limit 1; 
+        END IF;
+        IF c_symbol = 'S' THEN 
+            CALL skip_do_move(table_card_color, table_card_symbol, c_code, c_color, c_id, player_turn);
+        ELSEIF c_symbol = 'R' THEN
+            CALL reverse_do_move(table_card_color, table_card_symbol, c_code, c_color, c_id, player_turn);
+        ELSE
+            CALL regular_do_move(table_card_color, c_code, c_color, c_id, player_turn, table_card_symbol, c_symbol);
+        END IF;
 END$$
 
 DROP PROCEDURE IF EXISTS `do_wildcard_move`$$
 CREATE DEFINER=`root`@`localhost` PROCEDURE `do_wildcard_move` (IN `c_code` VARCHAR(3), IN `c_color` ENUM('R','Y','B','G'))  BEGIN
 	DECLARE c_id TINYINT;
 	DECLARE player_turn ENUM('p1', 'p2');
+    DECLARE c_symbol ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N');
+    DECLARE t_symbol ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N');
+    DECLARE plus SMALLINT;
+    SELECT plus_counter INTO plus FROM game_status LIMIT 1;
     SELECT p_turn INTO player_turn FROM game_status;
-    SELECT h.card_id into c_id FROM hand h INNER JOIN deck d on h.card_id=d.card_id WHERE d.card_code=c_code AND h.player_name=player_turn LIMIT 1;    
+    SELECT h.card_id, d.card_symbol INTO c_id, c_symbol FROM hand h INNER JOIN deck d ON h.card_id=d.card_id WHERE d.card_code=c_code AND h.player_name=player_turn LIMIT 1;
+    SELECT d.card_symbol INTO t_symbol FROM table_deck t INNER JOIN deck d ON d.card_code=t.card_code ORDER BY table_id DESC LIMIT 1;
     IF c_color != '' THEN
-        INSERT INTO table_deck (card_code, wild_card_color) VALUES (c_code, c_color);
-        DELETE FROM hand WHERE player_name=player_turn AND card_id=c_id;
-        IF player_turn = 'p1' THEN
-            UPDATE game_status SET p_turn='p2', draw_counter=0; 
-        ELSE 
-            UPDATE game_status SET p_turn='p1', draw_counter=0; 
-        END IF;
+    	IF (t_symbol = '+4' OR t_symbol = '+2') AND plus != 0 THEN
+            IF c_symbol = '+4' THEN
+                CALL plus_counter(c_symbol);
+                INSERT INTO table_deck (card_code, p_name, wild_card_color) VALUES (c_code, player_turn, c_color);
+                DELETE FROM hand WHERE player_name=player_turn AND card_id=c_id;
+                IF player_turn = 'p1' THEN
+                    UPDATE game_status SET p_turn='p2', draw_counter=0; 
+                ELSE 
+                    UPDATE game_status SET p_turn='p1', draw_counter=0; 
+                END IF;
+			END IF;
+        ELSE
+        	IF c_symbol = '+4' THEN
+                CALL plus_counter(c_symbol);
+            END IF;
+        	INSERT INTO table_deck (card_code, p_name, wild_card_color) VALUES (c_code, player_turn, c_color);
+            DELETE FROM hand WHERE player_name=player_turn AND card_id=c_id;
+            IF player_turn = 'p1' THEN
+            	UPDATE game_status SET p_turn='p2', draw_counter=0; 
+            ELSE 
+            	UPDATE game_status SET p_turn='p1', draw_counter=0; 
+            END IF;
+		END IF;           
 	END IF;
 END$$
 
@@ -99,22 +134,64 @@ DROP PROCEDURE IF EXISTS `pass`$$
 CREATE DEFINER=`root`@`localhost` PROCEDURE `pass` ()  BEGIN
 DECLARE player_turn ENUM('p1', 'p2');
 DECLARE d_counter smallint;
-SELECT draw_counter INTO d_counter FROM game_status LIMIT 1;
-SELECT p_turn, draw_counter INTO player_turn, d_counter FROM game_status LIMIT 1;
+DECLARE p_counter smallint;
+SELECT p_turn, draw_counter, plus_counter INTO player_turn, d_counter, p_counter FROM game_status LIMIT 1;
 IF d_counter = 1 THEN
+	IF p_counter > 0 THEN 
+    	loop_label: LOOP
+        	IF p_counter = 0 THEN
+            	LEAVE loop_label;
+            END IF;
+            SET p_counter = p_counter - 1;
+            CALL general_draw(player_turn);
+		END LOOP;
+    END IF;
     IF player_turn = 'p1' THEN
-        UPDATE game_status SET p_turn='p2', draw_counter = 0;
+        UPDATE game_status SET p_turn='p2', draw_counter = 0, plus_counter = 0;
     ELSEIF player_turn = 'p2' THEN
-        UPDATE game_status SET p_turn='p1', draw_counter =0;
+        UPDATE game_status SET p_turn='p1', draw_counter =0, plus_counter = 0;
     END IF;
 END IF;
 END$$
 
+DROP PROCEDURE IF EXISTS `plus_counter`$$
+CREATE DEFINER=`root`@`localhost` PROCEDURE `plus_counter` (IN `c_symbol` ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N'))  BEGIN 
+DECLARE counter SMALLINT;
+SELECT plus_counter INTO counter FROM game_status;
+IF c_symbol = '+4' THEN
+	SET counter = counter + 4;
+ELSEIF c_symbol = '+2' THEN
+	SET counter = counter + 2;
+END IF;
+UPDATE game_status SET plus_counter = counter;
+END$$
+
+DROP PROCEDURE IF EXISTS `plus_do_move`$$
+CREATE DEFINER=`root`@`localhost` PROCEDURE `plus_do_move` (IN `table_card_color` ENUM('R','Y','B','G','W'), IN `table_card_symbol` ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N'), IN `c_code` VARCHAR(3))  BEGIN
+DECLARE c_symbol ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N');
+DECLARE c_color ENUM('R','Y','B','G','W');
+SELECT d.card_color, d.card_symbol into c_color, c_symbol FROM hand h INNER JOIN deck d on h.card_id=d.card_id WHERE d.card_code=c_code AND h.player_name=player_turn LIMIT 1;
+
+IF table_card_symbol = '+4' THEN
+	SELECT wild_card_color INTO table_card_color FROM table_deck order by table_id desc limit 1;
+    IF c_color = table_card_color AND c_symbol = '+2' THEN
+    	CALL do_move(table_card_color, table_card_symbol, c_code);
+    END IF;
+ELSEIF table_card_symbol = '+2' THEN 
+    IF c_symbol = '+2' THEN
+    	CALL do_move(table_card_color, table_card_symbol, c_code);
+	END IF;
+END IF;    	
+	
+END$$
+
 DROP PROCEDURE IF EXISTS `regular_do_move`$$
-CREATE DEFINER=`root`@`localhost` PROCEDURE `regular_do_move` (IN `t_color` ENUM('R','Y','B','G','W'), IN `c_code` VARCHAR(3), IN `c_color` ENUM('R','Y','B','G','W'), IN `c_id` TINYINT, IN `player_turn` ENUM('p1','p2'), IN `t_symbol` ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N'), IN `c_symbol` ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N'))  NO SQL
-BEGIN
+CREATE DEFINER=`root`@`localhost` PROCEDURE `regular_do_move` (IN `t_color` ENUM('R','Y','B','G','W'), IN `c_code` VARCHAR(3), IN `c_color` ENUM('R','Y','B','G','W'), IN `c_id` TINYINT, IN `player_turn` ENUM('p1','p2'), IN `t_symbol` ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N'), IN `c_symbol` ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N'))  BEGIN
 	IF c_color = t_color OR c_symbol = t_symbol THEN
-    	INSERT INTO table_deck (card_code) VALUES (c_code);
+    	IF c_symbol = '+2' THEN
+        	CALL plus_counter(c_symbol);
+        END IF;
+    	INSERT INTO table_deck (card_code, p_name) VALUES (c_code, player_turn);
         DELETE FROM hand WHERE player_name=player_turn AND card_id=c_id;
         IF player_turn = 'p1' THEN
         	UPDATE game_status SET p_turn='p2', draw_counter=0; 
@@ -125,10 +202,9 @@ BEGIN
 END$$
 
 DROP PROCEDURE IF EXISTS `reverse_do_move`$$
-CREATE DEFINER=`root`@`localhost` PROCEDURE `reverse_do_move` (IN `t_color` ENUM('R','Y','B','G','W'), IN `t_symbol` ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N'), IN `c_code` VARCHAR(3), IN `c_color` ENUM('R','Y','B','G','W'), IN `c_id` TINYINT, IN `player_turn` ENUM('p1','p2'))  NO SQL
-BEGIN
+CREATE DEFINER=`root`@`localhost` PROCEDURE `reverse_do_move` (IN `t_color` ENUM('R','Y','B','G','W'), IN `t_symbol` ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N'), IN `c_code` VARCHAR(3), IN `c_color` ENUM('R','Y','B','G','W'), IN `c_id` TINYINT, IN `player_turn` ENUM('p1','p2'))  BEGIN
 	IF t_symbol='R' OR t_color=c_color THEN
-        INSERT INTO table_deck (card_code) VALUES (c_code);
+        INSERT INTO table_deck (card_code, p_name) VALUES (c_code, player_turn);
         DELETE FROM hand WHERE player_name=player_turn AND card_id=c_id;
         UPDATE game_status SET p_turn=player_turn; 
 	END IF;
@@ -137,7 +213,7 @@ END$$
 DROP PROCEDURE IF EXISTS `skip_do_move`$$
 CREATE DEFINER=`root`@`localhost` PROCEDURE `skip_do_move` (IN `t_color` ENUM('R','Y','B','G','W'), IN `t_symbol` ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N'), IN `c_code` VARCHAR(3), IN `c_color` ENUM('R','Y','B','G','W'), IN `c_id` TINYINT, IN `player_turn` ENUM('p1','p2'))  BEGIN
 	IF t_symbol='S' OR t_color=c_color THEN
-        INSERT INTO table_deck (card_code) VALUES (c_code);
+        INSERT INTO table_deck (card_code, p_name) VALUES (c_code, player_turn);
         DELETE FROM hand WHERE player_name=player_turn AND card_id=c_id;
         UPDATE game_status SET p_turn=player_turn; 
 	END IF;
@@ -153,8 +229,9 @@ CREATE DEFINER=`root`@`localhost` PROCEDURE `start_game` ()  BEGIN
 DECLARE c_code varchar(3);
 DECLARE c_color ENUM('R', 'Y', 'B', 'G', 'W');
 DECLARE p_name ENUM('p1', 'p2');
-SELECT card_code, card_color INTO c_code, c_color FROM remaining_deck ORDER BY RAND() LIMIT 1;
-WHILE c_color = 'W' DO 
+DECLARE c_symbol ENUM('0','1','2','3','4','5','6','7','8','9','+4','+2','R','S','N');
+SELECT card_code, card_color, card_symbol INTO c_code, c_color, c_symbol FROM remaining_deck ORDER BY RAND() LIMIT 1;
+WHILE c_color = 'W' OR c_symbol = '+2' DO 
 	SELECT card_code, card_color INTO c_code, c_color FROM remaining_deck ORDER BY RAND() LIMIT 1;
 END WHILE;
 ALTER TABLE table_deck AUTO_INCREMENT = 0;
@@ -164,8 +241,7 @@ DELETE FROM remaining_deck WHERE c_code = card_code;
 END$$
 
 DROP PROCEDURE IF EXISTS `uno_status`$$
-CREATE DEFINER=`root`@`localhost` PROCEDURE `uno_status` ()  NO SQL
-BEGIN 
+CREATE DEFINER=`root`@`localhost` PROCEDURE `uno_status` ()  BEGIN 
 	DECLARE usern varchar(20);
     DECLARE uno ENUM('active', 'not active');
 	SELECT p.username, p.uno_status INTO usern, uno FROM game_status g inner join player p on p.player_name=g.p_turn;
@@ -317,6 +393,7 @@ CREATE TABLE `game_status` (
   `status` enum('not active','initialized','started','ended','aborded') COLLATE utf8_bin NOT NULL DEFAULT 'not active',
   `p_turn` enum('p1','p2') COLLATE utf8_bin DEFAULT NULL,
   `draw_counter` smallint(6) NOT NULL DEFAULT 0,
+  `plus_counter` smallint(6) NOT NULL DEFAULT 0,
   `result` enum('p1','p2','d') COLLATE utf8_bin DEFAULT NULL,
   `last_change` timestamp NOT NULL DEFAULT current_timestamp() ON UPDATE current_timestamp()
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin;
@@ -325,8 +402,8 @@ CREATE TABLE `game_status` (
 -- Άδειασμα δεδομένων του πίνακα `game_status`
 --
 
-INSERT INTO `game_status` (`status`, `p_turn`, `draw_counter`, `result`, `last_change`) VALUES
-('started', 'p1', 0, NULL, '2019-12-24 15:46:06');
+INSERT INTO `game_status` (`status`, `p_turn`, `draw_counter`, `plus_counter`, `result`, `last_change`) VALUES
+('not active', NULL, 0, 0, NULL, '2019-12-25 16:06:50');
 
 --
 -- Δείκτες `game_status`
@@ -356,12 +433,20 @@ CREATE TABLE `hand` (
 --
 
 INSERT INTO `hand` (`player_name`, `card_id`) VALUES
-('p1', 33),
+('p1', 9),
+('p1', 17),
+('p2', 18),
+('p2', 19),
+('p2', 21),
+('p1', 36),
 ('p1', 38),
-('p2', 47),
-('p1', 50),
+('p1', 40),
+('p2', 45),
+('p2', 60),
 ('p1', 63),
-('p2', 92);
+('p1', 70),
+('p2', 98),
+('p2', 100);
 
 -- --------------------------------------------------------
 
@@ -383,8 +468,8 @@ CREATE TABLE `player` (
 --
 
 INSERT INTO `player` (`player_name`, `username`, `uno_status`, `token`, `last_action`) VALUES
-('p1', 'FDSGFDG', 'not active', '72453f68ca58d5463c08802ccf528c56', '2019-12-24 15:41:30'),
-('p2', 'DFGFDG', 'not active', '93e15facdc83be34473b81cb08212c7f', '2019-12-24 15:41:34');
+('p1', NULL, 'not active', NULL, '2019-12-25 16:06:50'),
+('p2', NULL, 'not active', NULL, '2019-12-25 16:06:50');
 
 -- --------------------------------------------------------
 
@@ -409,18 +494,19 @@ INSERT INTO `remaining_deck` (`card_id`, `card_symbol`, `card_color`, `card_code
 (2, '1', 'R', '1R'),
 (3, '1', 'R', '1R'),
 (4, '2', 'R', '2R'),
+(5, '2', 'R', '2R'),
 (6, '3', 'R', '3R'),
 (7, '3', 'R', '3R'),
 (8, '4', 'R', '4R'),
-(9, '4', 'R', '4R'),
+(10, '5', 'R', '5R'),
 (11, '5', 'R', '5R'),
 (12, '6', 'R', '6R'),
 (13, '6', 'R', '6R'),
+(14, '7', 'R', '7R'),
 (15, '7', 'R', '7R'),
-(18, '9', 'R', '9R'),
-(21, '+2', 'R', '+2R'),
-(22, 'R', 'R', 'RR'),
-(23, 'R', 'R', 'RR'),
+(16, '8', 'R', '8R'),
+(20, '+2', 'R', '+2R'),
+(24, 'S', 'R', 'SR'),
 (25, 'S', 'R', 'SR'),
 (26, '0', 'Y', '0Y'),
 (27, '1', 'Y', '1Y'),
@@ -429,37 +515,45 @@ INSERT INTO `remaining_deck` (`card_id`, `card_symbol`, `card_color`, `card_code
 (30, '2', 'Y', '2Y'),
 (31, '3', 'Y', '3Y'),
 (32, '3', 'Y', '3Y'),
+(33, '4', 'Y', '4Y'),
 (34, '4', 'Y', '4Y'),
 (35, '5', 'Y', '5Y'),
-(36, '5', 'Y', '5Y'),
 (37, '6', 'Y', '6Y'),
 (39, '7', 'Y', '7Y'),
-(40, '7', 'Y', '7Y'),
 (41, '8', 'Y', '8Y'),
 (42, '8', 'Y', '8Y'),
 (43, '9', 'Y', '9Y'),
 (44, '9', 'Y', '9Y'),
-(45, '+2', 'Y', '+2Y'),
 (46, '+2', 'Y', '+2Y'),
+(47, 'R', 'Y', 'RY'),
 (48, 'R', 'Y', 'RY'),
 (49, 'S', 'Y', 'SY'),
+(50, 'S', 'Y', 'SY'),
 (51, '0', 'B', '0B'),
+(52, '1', 'B', '1B'),
 (53, '1', 'B', '1B'),
+(54, '2', 'B', '2B'),
+(55, '2', 'B', '2B'),
+(56, '3', 'B', '3B'),
+(57, '3', 'B', '3B'),
 (58, '4', 'B', '4B'),
 (59, '4', 'B', '4B'),
-(60, '5', 'B', '5B'),
 (61, '5', 'B', '5B'),
+(62, '6', 'B', '6B'),
 (64, '7', 'B', '7B'),
 (65, '7', 'B', '7B'),
 (66, '8', 'B', '8B'),
+(67, '8', 'B', '8B'),
 (68, '9', 'B', '9B'),
 (69, '9', 'B', '9B'),
-(70, '+2', 'B', '+2B'),
 (71, '+2', 'B', '+2B'),
 (72, 'R', 'B', 'RB'),
+(73, 'R', 'B', 'RB'),
 (74, 'S', 'B', 'SB'),
+(75, 'S', 'B', 'SB'),
 (76, '0', 'G', '0G'),
 (77, '1', 'G', '1G'),
+(78, '1', 'G', '1G'),
 (79, '2', 'G', '2G'),
 (80, '2', 'G', '2G'),
 (81, '3', 'G', '3G'),
@@ -467,16 +561,20 @@ INSERT INTO `remaining_deck` (`card_id`, `card_symbol`, `card_color`, `card_code
 (83, '4', 'G', '4G'),
 (84, '4', 'G', '4G'),
 (85, '5', 'G', '5G'),
+(86, '5', 'G', '5G'),
+(87, '6', 'G', '6G'),
 (88, '6', 'G', '6G'),
+(89, '7', 'G', '7G'),
 (90, '7', 'G', '7G'),
 (91, '8', 'G', '8G'),
+(92, '8', 'G', '8G'),
 (93, '9', 'G', '9G'),
+(94, '9', 'G', '9G'),
 (95, '+2', 'G', '+2G'),
 (96, '+2', 'G', '+2G'),
 (97, 'R', 'G', 'RG'),
-(98, 'R', 'G', 'RG'),
 (99, 'S', 'G', 'SG'),
-(100, 'S', 'G', 'SG'),
+(101, '+4', 'W', '4W'),
 (102, '+4', 'W', '4W'),
 (103, '+4', 'W', '4W'),
 (104, '+4', 'W', '4W'),
@@ -495,6 +593,7 @@ DROP TABLE IF EXISTS `table_deck`;
 CREATE TABLE `table_deck` (
   `table_id` tinyint(4) NOT NULL,
   `card_code` varchar(3) COLLATE utf8_bin NOT NULL,
+  `p_name` enum('p1','p2') COLLATE utf8_bin DEFAULT NULL,
   `wild_card_color` enum('R','Y','B','G') COLLATE utf8_bin DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin;
 
@@ -502,29 +601,8 @@ CREATE TABLE `table_deck` (
 -- Άδειασμα δεδομένων του πίνακα `table_deck`
 --
 
-INSERT INTO `table_deck` (`table_id`, `card_code`, `wild_card_color`) VALUES
-(1, '8R', NULL),
-(2, '2R', NULL),
-(3, '2B', NULL),
-(4, '1B', NULL),
-(5, 'RB', NULL),
-(6, '8B', NULL),
-(7, 'SB', NULL),
-(8, '3B', NULL),
-(9, '3B', NULL),
-(10, '2B', NULL),
-(11, '6B', NULL),
-(12, '4W', 'G'),
-(13, '1G', NULL),
-(14, '6G', NULL),
-(15, '5G', NULL),
-(16, '7G', NULL),
-(17, '9G', NULL),
-(18, '9R', NULL),
-(19, 'SR', NULL),
-(20, '5R', NULL),
-(21, '+2R', NULL),
-(22, '7R', NULL);
+INSERT INTO `table_deck` (`table_id`, `card_code`, `p_name`, `wild_card_color`) VALUES
+(1, 'RR', NULL, NULL);
 
 --
 -- Ευρετήρια για άχρηστους πίνακες
@@ -569,7 +647,7 @@ ALTER TABLE `table_deck`
 -- AUTO_INCREMENT για πίνακα `table_deck`
 --
 ALTER TABLE `table_deck`
-  MODIFY `table_id` tinyint(4) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=23;
+  MODIFY `table_id` tinyint(4) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=2;
 
 --
 -- Περιορισμοί για άχρηστους πίνακες
